+++ src/Events/ForceLogout.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace HasinHayder\TyroLogin\Events;
+
+class ForceLogout {
+    public function __construct(
+        public readonly int $userId
+    ) {}
+}